# reuse sort code and improve performance