--- a/pkg/vm/engine/tae/mergesort/types.go
+++ b/pkg/vm/engine/tae/mergesort/types.go
@@ -42,7 +42,6 @@ const nullFirst = true
 type SortElem[T any] struct {
 	data   T
 	isNull bool
-	idx    int32
 }
 
 type SortSlice[T any] struct {