## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [x] Code Refactoring

## Which issue(s) this PR fixes:

issue #15564

## What this PR does / why we need it:

improve sort performance.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
</head>

<body link="#0563C1" vlink="#954F72">


DATA   TYPE | new implementation | old implementation
-- | -- | --
TINYINT | 1.756814ms | 3.203524ms
SMALLINT | 2.942222ms | 4.420434ms
INT | 2.980723ms | 6.60475ms
BIGINT | 2.385118ms | 3.489127ms
TINYINT UNSIGNED | 1.665313ms | 2.134416ms
SMALLINT UNSIGNED | 2.800521ms | 3.173024ms
INT UNSIGNED | 3.153623ms | 4.836737ms
BIGINT UNSIGNED | 3.794229ms | 5.22284ms
FLOAT | 3.684928ms | 3.277525ms
DOUBLE | 7.167355ms | 4.750136ms
DATE | 324.702µs | 3.987031ms
DATETIME | 3.499527ms | 2.755721ms
VARCHAR | 8.386763ms | 6.950253ms
CHAR | 8.669665ms | 5.36664ms
BOOL | 590.804µs | 891.007µs
TIMESTAMP | 324.702µs | 3.825429ms
DECIMAL64 | 719.605µs | 2.317618ms
DECIMAL128 | 796.606µs | 2.61072ms
BINARY | 8.148162ms | 5.342141ms
VARBINARY | 7.750159ms | 7.683058ms
ENUM | 291.402µs | 2.440519ms
VECF32 | 3.271525ms | 6.455349ms
VECF64 | 1.450011ms | 9.19177ms
BIT | 7.600658ms | 3.476326ms
</body>
</html>
Benchmark result:
<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
</head>

<body link="#0563C1" vlink="#954F72">


test | new | old 
-- | -- | --
BenchmarkNewSortBlockColumns-8 | 631	   2047328 ns/op | 308	   3994324 ns/op
BenchmarkNewSortBlockColumns-8 | 553	   2832574 ns/op | 296	   7357301 ns/op



</body>

</html>